pyproject.toml
[project]
name = "dopejobs"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cdx-toolkit>=0.9.38",
    "html2text>=2025.4.15",
    "lxml>=6.0.2",
    "meilisearch>=0.40.0",
    "mlx-lm>=0.31.1",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
